+++ webforum/app/Http/Controllers/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -25,17 +26,26 @@ class PostController extends Controller
 
     public function store(Request $request)
     {
+    $user_id = Auth::user()->id;    
 
     $validator = Validator::make($request->all(), [
         'name' => 'required',
         'text' => 'required',
-        'user_id' => 'required',
         'thread_id' => 'required',
 
     ]);
 
     if ($validator->fails()) {
         return response()->json($validator->errors());
+    }
+
+    //MODERATOR
+    $isModerator = Auth::user()->isModerator;
+    //ADMIN
+    $isAdmin = Auth::user()->isAdmin;
+
+    if ($isModerator || $isAdmin) {
+        return response()->json(['error' => 'Unauthorized: Administrator or moderator cant create a post!!!'], 403);
     }
 
 
@@ -45,8 +55,9 @@ class PostController extends Controller
     $post->dateOfCreation = Carbon::now()->format('Y-m-d');
     $post->numberOfLikes = 0;
     $post->numberOfDislikes = 0;
-    $post->user_id = $request->user_id;
+    $post->user_id = $user_id;
     $post->thread_id = $request->thread_id;
+    $post->status = 'Posted';
 
     $post->save();
 
@@ -59,7 +70,6 @@ class PostController extends Controller
         $validator = Validator::make($request->all(), [
             'name' => 'required',
             'text' => 'required',
-            'user_id' => 'required',
             'thread_id' => 'required',
         ]);
 
@@ -68,15 +78,27 @@ class PostController extends Controller
             return response()->json($errors);
         }
 
+            //MODERATOR
+            $isModerator = Auth::user()->isModerator;
+            //ADMIN
+            $isAdmin = Auth::user()->isAdmin;
+
+            if ($isModerator || $isAdmin) {
+                return response()->json(['error' => 'Unauthorized: Administrator or moderator cant alter a post!!!'], 403);
+            }
+
         $post = Post::findOrFail($id);
+        $user_id = Auth::user()->id; 
+        $post_user_id = Post::where('id', $id)->value('user_id');
 
         $post->name = $request->name;
         $post->text = $request->text;
         $post->dateOfCreation = Carbon::now()->format('Y-m-d');
         $post->numberOfLikes = 0;
         $post->numberOfDislikes = 0;
-        $post->user_id = $request->user_id;
+        $post->user_id = $post_user_id;
         $post->thread_id = $request->thread_id;
+        $post->status = 'Posted';
 
         $post->save();
 
@@ -90,6 +112,13 @@ class PostController extends Controller
              'status' => 'required'
          ]);
 
+         //MODERATOR
+         $isModerator = Auth::user()->isModerator;
+
+         if (!$isModerator) {
+             return response()->json(['error' => 'Unauthorized: This is the function that can only be done by an administrator or by the user that created this comment!!!'], 403);
+         }
+
          $post = Post::findOrFail($id);
 
          $post->update(['status' => $request->input('status')]);
@@ -101,6 +130,12 @@ class PostController extends Controller
 
     public function destroy($id)
     {
+        //MODERATOR
+        $isModerator = Auth::user()->isModerator;
+
+        if (!$isModerator) {
+            return response()->json(['error' => 'Unauthorized: This is the function that can only be done by an administrator or by the user that created this post!!!'], 403);
+        }
         $post = Post::findOrFail($id);
         $post->delete();
         return response()->json('You have successfuly deleted a post!');
