+++ webforum/app/Http/Controllers/ThreadController.php
@@ -8,6 +8,8 @@ use App\Models\Thread;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Validator;
 
+use Illuminate\Support\Facades\Auth;
+
 class ThreadController extends Controller
 {
     public function index()
@@ -25,11 +27,11 @@ class ThreadController extends Controller
 
     public function store(Request $request)
     {
+    $user_id = Auth::user()->id; 
 
     $validator = Validator::make($request->all(), [
         'name' => 'required',
         'description' => 'required',
-        'user_id' => 'required',
 
     ]);
 
@@ -37,12 +39,21 @@ class ThreadController extends Controller
         return response()->json($validator->errors());
     }
 
+        //MODERATOR
+        $isModerator = Auth::user()->isModerator;
+        //ADMIN
+        $isAdmin = Auth::user()->isAdmin;
+    
+        if ($isModerator || $isAdmin) {
+            return response()->json(['error' => 'Unauthorized: Administrator or moderator cant create a thread!!!'], 403);
+        }
+
 
     $thread = new Thread();
     $thread->name = $request->name;
     $thread->description = $request->description;
     $thread->dateOfCreation = Carbon::now()->format('Y-m-d');
-    $thread->user_id = $request->user_id;
+    $thread->user_id = $user_id;
 
     $thread->save();
 
@@ -52,10 +63,17 @@ class ThreadController extends Controller
 
     public function update(Request $request, $id)
     {
+        //ADMIN
+        $isAdmin = Auth::user()->isAdmin;
+
+        if (!$isAdmin) {
+            return response()->json(['error' => 'Unauthorized: This is the function that can only be done by an administrator or by the user who created this thread!!!'], 403);
+        }
+
+
         $validator = Validator::make($request->all(), [
             'name' => 'required',
             'description' => 'required',
-            'user_id' => 'required',
         ]);
 
         if ($validator->fails()) {
@@ -64,11 +82,13 @@ class ThreadController extends Controller
         }
 
         $thread = Thread::findOrFail($id);
+        $user_id = Auth::user()->id; 
+        $thread_user_id = Thread::where('id', $id)->value('user_id');
 
         $thread->name = $request->name;
         $thread->description = $request->description;
         $thread->dateOfCreation = Carbon::now()->format('Y-m-d');
-        $thread->user_id = $request->user_id;
+        $thread->user_id = $thread_user_id;
 
         $thread->save();
 
@@ -76,23 +96,15 @@ class ThreadController extends Controller
     }
 
 
-    public function updateDescription(Request $request, $id)
-     {
-         $request->validate([
-             'description' => 'required'
-         ]);
-
-         $thread = Thread::findOrFail($id);
-
-         $thread->update(['description' => $request->input('description')]);
-
-         return response()->json(['message' => 'Thread description successfully altered.', new ThreadResource($thread)]);
-     }
-
-
-
     public function destroy($id)
     {
+        //ADMIN
+        $isAdmin = Auth::user()->isAdmin;
+
+        if (!$isAdmin) {
+            return response()->json(['error' => 'Unauthorized: This is the function that can only be done by an administrator or by the user who created this thread!!!'], 403);
+        }
+
         $thread = Thread::findOrFail($id);
         $thread->delete();
         return response()->json('You have successfuly deleted a thread!');
